+++ client/src/components/Header/index.jsx
@@ -1,0 +1,10 @@
+// Display the Page Header
+import Navbar from './Navbar'
+
+export default function Header() {
+  return (
+    <header>
+          <Navbar/>
+    </header>
+  );
+}