+++ client/src/components/Images/Images.jsx
@@ -1,0 +1,5 @@
+const Images = {
+    logo : new URL('./observatory.svg, import.meta.url).href')
+}
+
+export default Images